Move gitgraph from modules to services layer
